Remove SeriesState (and HighlightState) to new `states` directory (first of many)
